refactor: Improved code and docs
